requirements.txt
python-dotenv~=0.19.2
requests~=2.26.0
beautifulsoup4~=4.10.0
selenium~=4.1.0
webdriver-manager
mysql-connector-python~=8.0.28